refactor: gpt prompt 관리 enum에서 constants로 수정
