+++ src/main/java/com/cliptripbe/feature/video/application/VideoService.java
@@ -3,17 +3,13 @@ package com.cliptripbe.feature.video.application;
 import com.cliptripbe.feature.user.domain.User;
 import com.cliptripbe.feature.video.api.dto.request.ExtractPlaceRequestDto;
 import com.cliptripbe.infrastructure.openai.application.ChatGPTService;
-import com.cliptripbe.infrastructure.openai.enums.PromptEnum;
+import com.cliptripbe.infrastructure.openai.prompt.PromptConstants;
 import com.cliptripbe.infrastructure.openai.utils.ChatGptUtils;
-import com.cliptripbe.infrastructure.youtube.YoutubeService;
-import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import reactor.core.publisher.Mono;
 import reactor.core.scheduler.Schedulers;
 
 @Slf4j
@@ -28,7 +24,7 @@ public class VideoService {
         String caption = "전체 자막";
 
         String requestPrompt =
-            caption + System.lineSeparator() + PromptEnum.EXTRACT_CAPTION.getPrompt();
+            caption + System.lineSeparator() + PromptConstants.EXTRACT_CAPTION;
 
         String rawResponse = chatGPTService.ask(requestPrompt)
             .subscribeOn(Schedulers.boundedElastic())
